fix(*): fix CBD tests
